Worked on drawFreeform(), added bezierHelper()
drawFreeform will be a function where you can input a simple table of coordinates and instructions and it will draw what you have mapped out. In the process of making this I started working on bezierHelper() which helps visualize how the bezier curve works.

Ideally I would integrate a helper layer that can be toggled into the drawFreeform function, but we'll see how realistic that is.