Restart project (#23)
* Add create_button.lua

I'm working here on the higher order tiling function as well. Need to experiment with how to get it to take more parameters in a manageable way.

* Update create_button.lua

Got the grid working with a higher-order function

* Add pad and radius options with error checking

* Add shrinkOnPress, border and refactor

Button can shrink on press now. Border is possible and the border goes inwards. Refactored the tileFn to remove unnecessary args. Now also possible to have a height separat from width.

* Create print.lua

Adds ability to print to a "console" under the Canvas node for debugging purposes.

* Delete deprecated code

* Add htmlColors, print, and beginnings of parseSVG

- htmlColors is all the named HTML colors in a dictionary of rgba values
- print hacks a little debugger that floats beneath the Canvas node
- working on parseSVG but the pattern match gets messed up by dashes in the middle of attributes like stroke-width

* Update parseSVG.lua

* Update htmlColors.lua

* Add documentation

- Most of the files are empty, but README.md is updated.
- Also add builtins

---------

Co-authored-by: Mark Boyd <10804241+markalanboyd@users.noreply.github.com>